changed groupname to csc340 to match everyone. Security working!!!
