+++ Progress/pom.xml
@@ -8,7 +8,7 @@
 		<version>3.0.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.teamTBA</groupId>
+	<groupId>com.csc340</groupId>
 	<artifactId>Progress</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>Progress</name>
@@ -17,19 +17,34 @@
 		<java.version>17</java.version>
 	</properties>
 	<dependencies>
+                                    <dependency>
+                                                      <groupId>org.springframework.boot</groupId>
+                                                      <artifactId>spring-boot-starter-actuator</artifactId>
+                                    </dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
 		</dependency>
+                                    <dependency>
+                                                      <groupId>org.springframework.boot</groupId>
+                                                      <artifactId>spring-boot-starter-security</artifactId>
+                                    </dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
+                                    <dependency>
+                                                      <groupId>org.springframework.boot</groupId>
+                                                      <artifactId>spring-boot-starter-validation</artifactId>
+                                    </dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-
+                                    <dependency>
+                                                      <groupId>org.thymeleaf.extras</groupId>
+                                                      <artifactId>thymeleaf-extras-springsecurity6</artifactId>
+                                    </dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-devtools</artifactId>
@@ -51,6 +66,11 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+                                    <dependency>
+                                                      <groupId>org.springframework.security</groupId>
+                                                      <artifactId>spring-security-test</artifactId>
+                                                      <scope>test</scope>
+                                    </dependency>
 	</dependencies>
 
 	<build>
