+++ Progress/src/main/java/com/csc340/Progress/ProgressApplication.java
@@ -1,0 +1,13 @@
+package com.csc340.Progress;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProgressApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProgressApplication.class, args);
+	}
+
+}
