+++ Progress/src/main/java/com/csc340/Progress/ad/AdRepository.java
@@ -1,0 +1,11 @@
+package com.csc340.Progress.ad;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ *
+ * @author Dallin Pierce
+ */
+public interface AdRepository extends JpaRepository<Ad, Long> {
+    
+}
