+++ Progress/src/main/java/com/csc340/Progress/gymUser/GymUser.java
@@ -1,0 +1,37 @@
+package com.csc340.Progress.gymUser;
+
+import java.sql.Date;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ *
+ * @author Dallin Pierce
+ */
+@AllArgsConstructor
+@Entity
+@Table(name = "user")
+@NoArgsConstructor
+@Getter
+@Setter
+public class GymUser {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String username;
+    private String firstName;
+    private String lastName;
+    private Date dob;
+    private String gender;
+    private int fitnessLevel;
+    private String password;
+    
+    
+}
