+++ Progress/src/main/java/com/csc340/Progress/video/VideoRepository.java
@@ -1,0 +1,12 @@
+package com.csc340.Progress.video;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ *
+ * @author Dallin Pierce
+ */
+public interface VideoRepository extends JpaRepository<Video, Long>{
+    public List<Video> findVideosByTrainerId(long trainerId);
+}
