--- Progress/src/main/java/com/teamTBA/Progress/ProgressApplication.java
@@ -1,13 +1,0 @@
-package com.teamTBA.Progress;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProgressApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProgressApplication.class, args);
-	}
-
-}
