--- Progress/src/main/java/com/teamTBA/Progress/ad/AdRepository.java
@@ -1,11 +1,0 @@
-package com.teamTBA.Progress.ad;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- *
- * @author Dallin Pierce
- */
-public interface AdRepository extends JpaRepository<Ad, Long> {
-    
-}
