--- Progress/src/main/java/com/teamTBA/Progress/user/User.java
@@ -1,37 +1,0 @@
-package com.teamTBA.Progress.user;
-
-import java.sql.Date;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- *
- * @author Dallin Pierce
- */
-@AllArgsConstructor
-@Entity
-@Table(name = "user")
-@NoArgsConstructor
-@Getter
-@Setter
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String username;
-    private String firstName;
-    private String lastName;
-    private Date dob;
-    private String gender;
-    private int fitnessLevel;
-    private String password;
-    
-    
-}
