--- Progress/src/main/java/com/teamTBA/Progress/user/UserRepository.java
@@ -1,11 +1,0 @@
-package com.teamTBA.Progress.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- *
- * @author Dallin Pierce
- */
-public interface UserRepository extends JpaRepository<User, Long> {
-    public User findUserById(long id);
-}
