--- Progress/src/main/java/com/teamTBA/Progress/video/VideoService.java
@@ -1,36 +1,0 @@
-package com.teamTBA.Progress.video;
-
-import com.teamTBA.Progress.user.UserRepository;
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-/**
- *
- * @author Dallin Pierce
- */
-@Service
-public class VideoService {
-    
-    @Autowired
-    private VideoRepository repo;
-    
-    @Autowired
-    private UserRepository userRepo;
-    
-    public List<Video> getVideosByTrainerId(long trainerId) {
-        List<Video> videoList = repo.findVideosByTrainerId(trainerId);
-        for(Video video : videoList){
-            video.setTrainer(userRepo.findUserById(video.getTrainerId()));
-        }
-        return videoList;
-    }
-    
-    public Video getVideo(long id) {
-        return repo.getReferenceById(id);
-    }
-    
-    void saveVideo(Video video) {
-        repo.save(video);
-    }
-}
