+++ Progress/src/main/resources/templates/video/post.html
@@ -10,7 +10,7 @@
 </head>
 <body>
     <div id="header">
-        <h2><a href="#">Progress</a></h2>
+        <h2><a th:href="@{/dashboard}">Progress</a></h2>
 
         <div id="menu">
