+++ Progress/src/main/resources/templates/video/videos.html
@@ -11,7 +11,7 @@
 <body>
     <div class="page-container">
         <header>
-            <h1><a href="index-trainer.html">Progress</a></h1>
+            <h1><a th:href="@{/dashboard}">Progress</a></h1>
             <div id="menu">
 
                 <div class="hamburger">
