+++ Progress/src/main/resources/templates/video/view_video.html
@@ -12,7 +12,7 @@
 <body>
     <div class="page-container">
         <header>
-            <h1><a href="index.html">Progress</a></h1>
+            <h1><a th:href="@{/dashboard}">Progress</a></h1>
             <div id="menu">
 
                 <div class="hamburger">
