+++ Progress/src/test/java/com/csc340/Progress/ProgressApplicationTests.java
@@ -1,0 +1,13 @@
+package com.csc340.Progress;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProgressApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
