--- Progress/src/test/java/com/teamTBA/Progress/ProgressApplicationTests.java
@@ -1,13 +1,0 @@
-package com.teamTBA.Progress;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProgressApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
